- implemented Rendering of Values and Factors

+++ src/main/kotlin/domain/entities/Entities.kt
@@ -119,5 +119,5 @@ data class EntityCountable<T : IEntity<*>> constructor(
 data class EntityValuable<T : IEntity<*>> constructor(
     val entity: T,
     val value: String?,
-    val factor: Float? = null
+    val factor: Int? = null
 )
